requirements.txt
streamlit==1.11.0
pandas==1.2.3
numpy==1.20.2
base58==2.1.0
ipython==7.22.0
psycopg2-binary==2.8.2
pendulum==2.1.2
matplotlib==3.6.0
